[Fix] main() 메서드에 move 명령 추가
1. MoveResult의 score값을 double로 수정
2. GamePlayService는 Board를 출력할 때 String으로 보여준다.

+++ src/main/java/pe/goblin/chess/application/usecase/GamePlayUseCase.java
@@ -6,6 +6,6 @@ import pe.goblin.chess.exception.ApplicationException;
 public interface GamePlayUseCase {
     MoveResult move(String input) throws ApplicationException;
 
-    record MoveResult(GameStatus gameStatus, Integer whiteScore, Integer blackScore, String board) {
+    record MoveResult(GameStatus gameStatus, double whiteScore, double blackScore, String board) {
     }
 }
